renameing files to be more findable and consistent
