+++ pkg/objectcache/containerprofilecache/removal_grace_test.go
@@ -12,23 +12,7 @@ import (
 	corev1 "k8s.io/api/core/v1"
 )
 
-// TestProjectedProfile_SurvivesContainerRemovalGrace pins the end-of-life
-// contract for the event pipeline: when a container is removed, events that
-// were emitted during its life are still in flight (ordered event queue 50ms
-// collection tick + batching + worker pool), and the rule engine resolves the
-// projected profile by container ID at evaluation time. Deleting the cache
-// entry immediately on the remove callback makes rules with
-// ProfileDependency=Required (e.g. R0001) silently suppress the container's
-// terminal events as "profile_incomplete".
-//
-// Evidence: CI run 31846699597 Test_48 — init container "setup"
-// (sh -c "sleep 75; /usr/bin/id"): remove processed at 22:44:26, terminal exec
-// evaluated afterwards, zero R0001 despite adopted profile and 98 R0003 during
-// the container's life.
-//
-// Contract: the projected profile must remain resolvable for a grace window
-// after the remove callback (long enough to cover the event pipeline delay),
-// and only then be evicted.
+// Projected profile must stay resolvable for a grace window after the remove callback, then be evicted — in-flight terminal events must not lose their profile (else ProfileDependency=Required rules silently suppress as profile_incomplete). Evidence: CI 31846699597 Test_48 init "setup" — 0 R0001 on terminal exec despite adopted profile.
 func TestProjectedProfile_SurvivesContainerRemovalGrace(t *testing.T) {
 	c, _ := newTestCache(t, &fakeProfileClient{})
