+++ pkg/rulemanager/cel/libraries/containerprofile/containerprofile.go
