+++ pkg/rulemanager/cel/libraries/containerprofilenetwork/containerprofilenetwork.go
