+++ tests/resources/percontainer-cp-sidecar.yaml
@@ -1,0 +1,16 @@
+# Authored ContainerProfile for the "sidecar" container of the percontainer pod, resolved
+# as "<label>-<containerName>" == "percontainer-sidecar". Allows whoami but not id (the
+# inverse of percontainer-app), so per-container binding is verifiable.
+apiVersion: spdx.softwarecomposition.kubescape.io/v1beta1
+kind: ContainerProfile
+metadata:
+  name: percontainer-sidecar
+spec:
+  execs:
+  - path: /usr/bin/sleep
+    args:
+    - /usr/bin/sleep
+    - infinity
+  - path: /usr/bin/whoami
+    args:
+    - /usr/bin/whoami
