+++ tests/resources/percontainer-deployment.yaml
@@ -1,23 +1,23 @@
-# 2-container pod bound via one pod label kubescape.io/user-defined-profile=mc35.
+# 2-container pod bound via one pod label kubescape.io/user-defined-profile=percontainer.
 # ContainerProfiles are per-container: node-agent resolves each container to
-# <label>-<containerName> (mc35-app / mc35-sidecar), falling back to the bare
+# <label>-<containerName> (percontainer-app / percontainer-sidecar), falling back to the bare
 # <label> for single-container pods.
 apiVersion: apps/v1
 kind: Deployment
 metadata:
-  name: mc35-deployment
+  name: percontainer-deployment
   labels:
-    app: mc35
+    app: percontainer
 spec:
   replicas: 1
   selector:
     matchLabels:
-      app: mc35
+      app: percontainer
   template:
     metadata:
       labels:
-        app: mc35
-        kubescape.io/user-defined-profile: mc35
+        app: percontainer
+        kubescape.io/user-defined-profile: percontainer
     spec:
       containers:
       - name: app
